+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Enhancement:
 
 - Added `Request-SddcManagerPasswordExpiration` cmdlet to retrieve the password expiration policy for the default local users on an SDDC Manager appliance. [GH-97](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/97)
 - Added `Update-SddcManagerPasswordExpiration` cmdlet to update the password expiration policy for the default local users on an SDDC Manager appliance. [GH-97](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/97)
+- Added `Request-PasswordRotationPolicy` cmdlet to retrieve the password rotation settings for accounts managed by SDDC Manager. Results can be filtered by resource type or workload domain. [GH-100](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/100)
+- Added `Publish-PasswordRotationPolicy` cmdlet to publish the password rotation settings for accounts managed by SDDC Manager in HTML or JSON format. Results can be filtered by resource type or workload domain. [GH-100](https://github.com/vmware/powershell-module-for-vmware-cloud-foundation-password-management/pull/100)
 
 Bug Fixes: